add lib/processing/image.rb
